rag/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
chromadb>=0.4.0
httpx>=0.25.0
python-dotenv>=1.0.0
pydantic>=2.0.0
llama-index-core>=0.10.0
llama-index-embeddings-huggingface>=0.2.0
llama-index-vector-stores-chroma>=0.1.0
sentence-transformers>=2.2.0
torch>=2.0.0
transformers>=4.35.0